+++ examples/desktop_private_vpc/variables.tf
@@ -16,4 +16,12 @@ variable "no_proxy" {}
 
 variable "key_name" {}
 
+variable "enable_cluster_autoscaling" {}
+
+variable "enable_pod_autoscaling" {}
+
+variable "cluster_version" {}
+
+variable "protect_cluster_from_scale_in" {}
+
 variable "desired_worker_nodes" {}