+++ examples/ec2_node_private_vpc/locals.tf
@@ -1,10 +1,10 @@
 locals {
   aws_authenticator_env_variables = {
-#    AWS_PROFILE = "${var.aws_profile}"
+    #    AWS_PROFILE = "${var.aws_profile}"
   }
 
   tags = {
-    Owner       = "Peter Parker"
-    Environment = "dev"
+    ProductDomain = "${var.cluster_prefix}"
+    Environment   = "dev"
   }
 }